Moved the command rules file to the top-level directory. Added a decorator to replace repeated code in command rule functions. Added a notimplemented decorator that raises a NotImplementedError, to add for commands that aren't implemented yet.
